+++ packages/angular/src/executors/package/schema.json
@@ -8,10 +8,6 @@
   "type": "object",
   "presets": [
     {
-      "name": "Publishable Library with Tailwind",
-      "keys": ["project", "tailwindConfig"]
-    },
-    {
       "name": "Updating Project Dependencies for Publishable Library",
       "keys": ["project"]
     }
@@ -36,12 +32,6 @@
     "poll": {
       "type": "number",
       "description": "Enable and define the file watching poll time period in milliseconds. _Note: this is only supported in Angular versions >= 18.0.0_."
-    },
-    "tailwindConfig": {
-      "type": "string",
-      "description": "The full path for the Tailwind configuration file, relative to the workspace root. If not provided and a `tailwind.config.js` file exists in the project or workspace root, it will be used. Otherwise, Tailwind will not be configured. _Note: starting with Angular v17, this option is no longer used and the configuration will be picked up if exists at the project or workspace root_.",
-      "x-completion-type": "file",
-      "x-completion-glob": "tailwind.config@(.js|.cjs|.mjs|.ts)"
     }
   },
   "additionalProperties": false,
