+++ packages/angular/src/generators/setup-tailwind/lib/index.ts
@@ -1,6 +1,6 @@
-export * from './add-tailwind-config-path-to-project';
 export * from './add-tailwind-config';
 export * from './add-tailwind-required-packages';
 export * from './detect-tailwind-installed-version';
 export * from './normalize-options';
 export * from './update-application-styles';
+export * from './validate-build-target';
