+++ packages/angular/src/generators/setup-tailwind/setup-tailwind.ts
@@ -6,11 +6,11 @@ import {
 } from '@nx/devkit';
 import {
   addTailwindConfig,
-  addTailwindConfigPathToProject,
   addTailwindRequiredPackages,
   detectTailwindInstalledVersion,
   normalizeOptions,
   updateApplicationStyles,
+  validateBuildTarget,
 } from './lib';
 import { GeneratorOptions } from './schema';
 
@@ -20,6 +20,12 @@ export async function setupTailwindGenerator(
 ): Promise<GeneratorCallback> {
   const options = normalizeOptions(rawOptions);
   const project = readProjectConfiguration(tree, options.project);
+
+  if (rawOptions.buildTarget && !project.targets?.[rawOptions.buildTarget]) {
+    throw new Error(
+      `The provided target "${options.buildTarget}" was not found for project "${options.project}". Please provide a valid build target.`
+    );
+  }
 
   const tailwindInstalledVersion = detectTailwindInstalledVersion(tree);
 
@@ -35,7 +41,7 @@ export async function setupTailwindGenerator(
   if (project.projectType === 'application') {
     updateApplicationStyles(tree, options, project);
   } else if (project.projectType === 'library') {
-    addTailwindConfigPathToProject(tree, options, project);
+    validateBuildTarget(options, project);
   }
 
   if (!options.skipFormat) {
